add order creation

+++ internal/app/domain/balance/balance.go
@@ -3,11 +3,11 @@ package balance
 //go:generate easyjson -all balance.go
 type Balance struct {
 	Current   float64 `json:"current"`
-	Withdrawn int     `json:"withdrawn"`
+	Withdrawn float64 `json:"withdrawn"`
 	UserID    string  `json:"-"`
 }
 
-func New(current float64, withdrawn int, userID string) Balance {
+func New(current float64, withdrawn float64, userID string) Balance {
 	return Balance{
 		Current:   current,
 		Withdrawn: withdrawn,
